+++ src/ResourceManager/Compute/Commands.Compute/Extension/AzureDiskEncryption/AzureDiskEncryptionExtensionConstants.cs
@@ -39,5 +39,7 @@ namespace Microsoft.Azure.Commands.Compute.Extension.AzureDiskEncryption
         public const string osTypeLinux = "Linux";
         public const string osTypeWindows = "Windows";
         public const string defaultKeyEncryptionAlgorithm = "RSA-OAEP";
+        public const string keyVaultResourceIdKey = "KeyVaultResourceId";
+        public const string kekVaultResourceIdKey = "KekVaultResourceId";
     }
 }